Exercise List 08 Done
Left a comment on Ex07, because it suggested to do the exercises on the book. Turned out it had 50 of them, so I just didn't do.
Besides, the subject of vectors doesn't seem to be hard or to have any particular problem or concept that I have to worke more on.